feat : Schedule은 오직 로그인된 유저들만 접근할 수 있도록 수정.
modify와 delete는 로그인된 유저와 수정/삭제하고자하는 schedule의 userId가 같은 경우에만 할 수 있도록 수정

로직 변경에 따른 테스트 코드 수정

+++ src/main/java/org/zerock/schedule_app_develop/service/ScheduleService.java
@@ -4,12 +4,14 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.zerock.schedule_app_develop.dto.LoginSessionAttribute;
 import org.zerock.schedule_app_develop.dto.ScheduleCreateRequestDto;
 import org.zerock.schedule_app_develop.dto.ScheduleResponseDto;
 import org.zerock.schedule_app_develop.dto.ScheduleUpdateRequestDto;
 import org.zerock.schedule_app_develop.entity.Schedule;
 import org.zerock.schedule_app_develop.entity.User;
 import org.zerock.schedule_app_develop.exception.ScheduleNotFoundException;
+import org.zerock.schedule_app_develop.exception.UnAuthorizedException;
 import org.zerock.schedule_app_develop.exception.UserNotFoundException;
 import org.zerock.schedule_app_develop.repository.ScheduleRepository;
 import org.zerock.schedule_app_develop.repository.UserRepository;
@@ -45,15 +47,22 @@ public class ScheduleService {
     }
 
     @Transactional
-    public ScheduleResponseDto updateSchedule(Long id, ScheduleUpdateRequestDto dto) {
+    public ScheduleResponseDto updateSchedule(Long id, ScheduleUpdateRequestDto dto, LoginSessionAttribute login) {
         Schedule schedule = scheduleRepository.findById(id).orElseThrow(() -> new ScheduleNotFoundException("Schedule not found"));
+        if (login.getId() != schedule.getUser().getId()) {
+            throw new UnAuthorizedException("권한 없음");
+        }
         schedule.modify(dto);
         schedule = scheduleRepository.saveAndFlush(schedule);
         return new ScheduleResponseDto(schedule);
     }
 
     @Transactional
-    public void deleteSchedule(Long id) {
+    public void deleteSchedule(Long id, LoginSessionAttribute login) {
+        Schedule schedule = scheduleRepository.findById(id).orElseThrow(() -> new ScheduleNotFoundException("Schedule not found"));
+        if(schedule.getUser().getId() != login.getId()) {
+            throw new UnAuthorizedException("권한 없음");
+        }
         scheduleRepository.deleteById(id);
     }
